Complete C# collection recovery and native ownership checks
Match widened collection sizing, repeated map headers and zero-bit message lists to the shared oracle. Keep retained unknown fields through damaged-body resets and count orphaned paths at save. Guard stale list erasure, read unaligned cook headers, use native u32 blob lengths and check pack extents before writes. Add builder fuzz adapters, allocator growth failures, descending allocation and pack sabotage regressions; document native ownership.

+++ internal/codegen/cstable/builder.go
@@ -138,7 +138,7 @@ public sealed unsafe class TableWorker
         if(length<0 || length>int.MaxValue) { return default; }
         byte* p=Raw(8+length+terminator);
         if(p==null) { return default; }
-        *(ulong*)p=(ulong)length;
+        *(uint*)p=(uint)length;
         return new TableBlobSlot { Pointer=(IntPtr)p,Length=length,ByteLength=length };
     }
 }
